requirements.txt
Flask
Pillow
numpy
gunicorn
requests